+++ lib/models/location.dart
@@ -8,12 +8,21 @@ class SelectedLocation {
 }
 
 class Location {
-  Location(this.id, this.displayName,
-      [this.grades, this.sections, this.climbs]);
+  Location(this.id, this.displayName, [this.gradesId, this.sections, this.climbs]);
 
   final String id;
   final String displayName;
-  final List<String> grades;
+  final String gradesId;
   final List<String> sections;
   final List<Climb> climbs;
+
+  Map<String, dynamic> get map {
+    return {
+      "id": id,
+      "displayName": displayName,
+      "grades": gradesId,
+      "sections": sections,
+      "climbs": climbs
+    };
+  }
 }