+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Firebase account creation and authentication
 - Retrieval of locations, climbs and attempts from firebase
 - Views for locations and climbs
+- Location creation